fix: export collisions
