+++ packages/server/src/lib.js
@@ -1,15 +1,7 @@
 export * from './api.js'
 export * from '@ucanto/core'
 export * from './server.js'
-export {
-  Failure,
-  MalformedCapability,
-  HandlerNotFound,
-  Link,
-  URI,
-  ok,
-  error,
-} from './server.js'
+export { Failure, Link, URI, ok, error } from './server.js'
 export {
   invoke,
   Invocation,
@@ -23,3 +15,4 @@ export { access, claim, Schema } from '@ucanto/validator'
 
 export * from './handler.js'
 export * as API from './api.js'
+export * as Error from './error.js'
